correction and inhancement by komal

+++ resources/views/admin/email_management/show_content.blade.php
@@ -84,7 +84,7 @@
                                                         </button>
                                                     </a>
 
-                                                    <a href='delete/{{ $user->id }}'>
+                                                    <a href="/admin/delete/{{ $user->id }}" onclick="return confirm('Are you sure you want to delete this email content ?')">
                                                         <button type="button" data-toggle="tooltip" title=""
                                                             class="btn btn-link btn-danger" data-original-title="Delete">
                                                             <i class="fa fa-times"></i>
